Remove everything from unnecessary closure in injection.ts
Originally, everything in injection.ts was wrapped in a closure so that none of the local variables and functions would be available to the console in the browser when using the extension.
However I failed to realise that the webpack build step automatically wraps everything in a closer anyway when compiling. Therefore there was never any need for the closure in the first place and it was just adding indentation to the file.
Also update webpack config to not copy over tsx files. Before it was only checking for .ts files.