Fixed difficult bug
Bug: Unchecked runtime.lastError: Could not establish connection. Receiving end does not exist.

Cause: The background script was accidentally importing a script which is meant only for the content script, which contained an immediate message request. Therefore it tried to send a message to itself as soon as it opened up.

How to debug: Run webpack with minification turned off and analyse the background.js generated file. It will list out all files that get concatenated and therefore you can deduce which one is not meant to be run in the background.